refactor: extract RecordTableRenderer + PlayerRecordSectionRenderer; wire into RecordHoldersView
