+++ ibl5/classes/RecordHolders/RecordHoldersView.php
@@ -26,35 +26,14 @@ use Security\HtmlSanitizer;
  */
 class RecordHoldersView implements RecordHoldersViewInterface
 {
-    /** @var array<string, string> */
-    private const STAT_LABELS = [
-        'Most Points' => 'Pts',
-        'Most Rebounds' => 'Reb',
-        'Most Assists' => 'Ast',
-        'Most Steals' => 'Stl',
-        'Most Blocks' => 'Blk',
-        'Most Turnovers' => 'TO',
-        'Most 3-Pointers' => '3PM',
-        'Most Three-Pointers' => '3PM',
-        'Most Minutes' => 'Min',
-        'Highest Scoring Average' => 'PPG',
-        'Highest Rebounding Average' => 'RPG',
-        'Highest Assists Average' => 'APG',
-        'Highest Steals Average' => 'SPG',
-        'Highest Blocks Average' => 'BPG',
-        'Highest 3-Point' => '3P%',
-        'Highest Three-Point' => '3P%',
-        'Highest Field Goal' => 'FG%',
-        'Highest Free Throw' => 'FT%',
-        'Best Season Record' => 'Record',
-        'Worst Season Record' => 'Record',
-        'Most Wins' => 'Wins',
-        'Most Losses' => 'Losses',
-        'Most Playoff Appearances' => 'Apps',
-        'Most Championship' => 'Titles',
-        'Longest Winning Streak' => 'Wins',
-        'Longest Losing Streak' => 'Losses',
-    ];
+    private RecordTableRenderer $tableRenderer;
+    private PlayerRecordSectionRenderer $playerSection;
+
+    public function __construct()
+    {
+        $this->tableRenderer = new RecordTableRenderer();
+        $this->playerSection = new PlayerRecordSectionRenderer($this->tableRenderer);
+    }
 
     /**
      * @see RecordHoldersViewInterface::render()
@@ -65,114 +44,12 @@ class RecordHoldersView implements RecordHoldersViewInterface
     {
         $output = '<h1 class="ibl-title">Record Holders</h1>';
         $output .= '<div class="record-section">';
-        $output .= $this->renderPlayerSingleGameRecords($records);
-        $output .= $this->renderPlayerFullSeasonRecords($records['playerFullSeason']);
-        $output .= $this->renderPlayerPlayoffRecords($records);
-        $output .= $this->renderPlayerHeatRecords($records);
+        $output .= $this->playerSection->renderPlayerSingleGameRecords($records);
+        $output .= $this->playerSection->renderPlayerFullSeasonRecords($records['playerFullSeason']);
+        $output .= $this->playerSection->renderPlayerPlayoffRecords($records);
+        $output .= $this->playerSection->renderPlayerHeatRecords($records);
         $output .= $this->renderTeamRecords($records);
         $output .= '</div>';
-
-        return $output;
-    }
-
-    // ---------------------------------------------------------------
-    // Section 1: Regular Season (Single Game)
-    // ---------------------------------------------------------------
-
-    /**
-     * Render Section 1: Player Regular Season (Single Game) records.
-     *
-     * @param AllRecordsData $records
-     */
-    private function renderPlayerSingleGameRecords(array $records): string
-    {
-        $output = '<div class="ibl-card">';
-        $output .= '<div class="ibl-card__header"><h2 class="ibl-card__title">Player, Regular Season (Single Game)</h2></div>';
-        $output .= '<div class="ibl-card__body">';
-
-        foreach ($records['playerSingleGame']['regularSeason'] as $category => $categoryRecords) {
-            $output .= $this->renderPlayerCategoryBlock($category, $categoryRecords);
-        }
-
-        // Quadruple Doubles
-        $output .= $this->renderPlayerCategoryBlock('Quadruple Doubles', $records['quadrupleDoubles'], true);
-
-        // Most All-Star Appearances
-        $output .= $this->renderAllStarBlock($records['allStarRecord']);
-
-        $output .= '</div></div>';
-
-        return $output;
-    }
-
-    // ---------------------------------------------------------------
-    // Section 2: Regular Season (Full Season)
-    // ---------------------------------------------------------------
-
-    /**
-     * Render Section 2: Player Regular Season (Full Season) records.
-     *
-     * @param array<string, list<FormattedSeasonRecord>> $seasonRecords
-     */
-    private function renderPlayerFullSeasonRecords(array $seasonRecords): string
-    {
-        $output = '<div class="ibl-card">';
-        $output .= '<div class="ibl-card__header"><h2 class="ibl-card__title">Player, Regular Season (Full Season) [minimum 50 games]</h2></div>';
-        $output .= '<div class="ibl-card__body">';
-
-        foreach ($seasonRecords as $category => $categoryRecords) {
-            $output .= $this->renderSeasonCategoryBlock($category, $categoryRecords);
-        }
-
-        $output .= '</div></div>';
-
-        return $output;
-    }
-
-    // ---------------------------------------------------------------
-    // Section 3: Playoffs
-    // ---------------------------------------------------------------
-
-    /**
-     * Render Section 3: Player Playoff records.
-     *
-     * @param AllRecordsData $records
-     */
-    private function renderPlayerPlayoffRecords(array $records): string
-    {
-        $output = '<div class="ibl-card">';
-        $output .= '<div class="ibl-card__header"><h2 class="ibl-card__title">Player, Playoffs</h2></div>';
-        $output .= '<div class="ibl-card__body">';
-
-        foreach ($records['playerSingleGame']['playoffs'] as $category => $categoryRecords) {
-            $output .= $this->renderPlayerCategoryBlock($category, $categoryRecords);
-        }
-
-        $output .= '</div></div>';
-
-        return $output;
-    }
-
-    // ---------------------------------------------------------------
-    // Section 4: H.E.A.T.
-    // ---------------------------------------------------------------
-
-    /**
-     * Render Section 4: Player H.E.A.T. records.
-     *
-     * @param AllRecordsData $records
-     */
-    private function renderPlayerHeatRecords(array $records): string
-    {
-        $output = '<div class="ibl-card">';
-        $output .= '<div class="ibl-card__header"><h2 class="ibl-card__title">Player, H.E.A.T.</h2></div>';
-        $output .= '<div class="ibl-card__body">';
-
-        foreach ($records['playerSingleGame']['heat'] as $category => $categoryRecords) {
-            $output .= $this->renderPlayerCategoryBlock($category, $categoryRecords);
-        }
-
-        $output .= '</div></div>';
 
         return $output;
     }
@@ -226,99 +103,20 @@ class RecordHoldersView implements RecordHoldersViewInterface
     // ---------------------------------------------------------------
 
     /**
-     * Render the shared category-block scaffold: a `.record-category` wrapper
-     * with a heading and a `.record-table` whose layout is fixed by a modifier
-     * class. Every per-category block differs only in that modifier, its
-     * colgroup, its header cells, and its pre-rendered body rows.
-     */
-    private function renderCategoryTable(string $category, string $modifierClass, string $colgroup, string $thead, string $rows): string
-    {
-        $output = '<div class="record-category">';
-        $output .= $this->renderCategoryHeading($category);
-        $output .= '<table class="ibl-data-table record-table ibl-table-subheading ' . $modifierClass . '" data-no-responsive>';
-        $output .= '<colgroup>' . $colgroup . '</colgroup>';
-        $output .= '<thead><tr>' . $thead . '</tr></thead>';
-        $output .= '<tbody>' . $rows . '</tbody></table></div>';
-
-        return $output;
-    }
-
-    /**
-     * Render a player single-game category block (heading + mini-table).
-     *
-     * @param list<FormattedPlayerRecord> $categoryRecords
-     */
-    private function renderPlayerCategoryBlock(string $category, array $categoryRecords, bool $multiLineAmount = false): string
-    {
-        $safeStatLabel = HtmlSanitizer::safeHtmlOutput($this->getStatColumnLabel($category));
-
-        $rows = '';
-        foreach ($categoryRecords as $record) {
-            $rows .= $this->renderPlayerRecordRow($record, $multiLineAmount);
-        }
-
-        return $this->renderCategoryTable(
-            $category,
-            'record-table--5col',
-            '<col class="col-player"><col class="col-team"><col class="col-date"><col class="col-opponent"><col class="col-amount">',
-            '<th>Player</th><th>Team</th><th>Date</th><th>Opponent</th><th>' . $safeStatLabel . '</th>',
-            $rows
-        );
-    }
-
-    /**
-     * Render a full-season category block (heading + mini-table).
-     *
-     * @param list<FormattedSeasonRecord> $categoryRecords
-     */
-    private function renderSeasonCategoryBlock(string $category, array $categoryRecords): string
-    {
-        $safeStatLabel = HtmlSanitizer::safeHtmlOutput($this->getStatColumnLabel($category));
-
-        $rows = '';
-        foreach ($categoryRecords as $record) {
-            $safeName = HtmlSanitizer::safeHtmlOutput($record['name']);
-            $safeTeam = HtmlSanitizer::safeHtmlOutput($record['teamAbbr']);
-            $safeSeason = HtmlSanitizer::safeHtmlOutput($record['season']);
-            $safeAmount = HtmlSanitizer::safeHtmlOutput($record['amount']);
-            $pid = $record['pid'];
-            $teamTid = $record['teamTid'];
-            $teamYr = (int) $record['teamYr'];
-
-            $seasonLink = '<a href="' . TeamCellHelper::teamPageUrl($teamTid, $teamYr) . '">' . $safeSeason . '</a>';
-
-            $rows .= '<tr>';
-            $rows .= PlayerImageHelper::renderLargePlayerCell($pid, $record['name']);
-            $rows .= '<td><a href="modules.php?name=Team&amp;op=team&amp;teamid=' . $teamTid . '&amp;yr=' . $teamYr . '"><img src="images/topics/' . $safeTeam . '.png" alt="' . strtoupper($safeTeam) . '"></a></td>';
-            $rows .= '<td>' . $seasonLink . '</td>';
-            $rows .= '<td class="ibl-stat-highlight">' . $safeAmount . '</td>';
-            $rows .= '</tr>';
-        }
-
-        return $this->renderCategoryTable(
-            $category,
-            'record-table--4col-season',
-            '<col class="col-player"><col class="col-team"><col class="col-season"><col class="col-amount">',
-            '<th>Player</th><th>Team</th><th>Season</th><th>' . $safeStatLabel . '</th>',
-            $rows
-        );
-    }
-
-    /**
      * Render a team game category block (heading + mini-table).
      *
      * @param list<FormattedTeamGameRecord> $categoryRecords
      */
     private function renderTeamGameCategoryBlock(string $category, array $categoryRecords): string
     {
-        $safeStatLabel = HtmlSanitizer::safeHtmlOutput($this->getStatColumnLabel($category));
+        $safeStatLabel = HtmlSanitizer::safeHtmlOutput($this->tableRenderer->getStatColumnLabel($category));
 
         $rows = '';
         foreach ($categoryRecords as $record) {
             $rows .= $this->renderTeamGameRow($record);
         }
 
-        return $this->renderCategoryTable(
+        return $this->tableRenderer->renderCategoryTable(
             $category,
             'record-table--4col-team',
             '<col class="col-team"><col class="col-date"><col class="col-opponent"><col class="col-amount">',
@@ -334,7 +132,7 @@ class RecordHoldersView implements RecordHoldersViewInterface
      */
     private function renderTeamSeasonCategoryBlock(string $category, array $categoryRecords): string
     {
-        $safeStatLabel = HtmlSanitizer::safeHtmlOutput($this->getStatColumnLabel($category));
+        $safeStatLabel = HtmlSanitizer::safeHtmlOutput($this->tableRenderer->getStatColumnLabel($category));
 
         $rows = '';
         foreach ($categoryRecords as $record) {
@@ -354,7 +152,7 @@ class RecordHoldersView implements RecordHoldersViewInterface
             $rows .= '</tr>';
         }
 
-        return $this->renderCategoryTable(
+        return $this->tableRenderer->renderCategoryTable(
             $category,
             'record-table--3col-team-season',
             '<col class="col-team"><col class="col-season"><col class="col-amount">',
@@ -370,7 +168,7 @@ class RecordHoldersView implements RecordHoldersViewInterface
      */
     private function renderFranchiseCategoryBlock(string $category, array $categoryRecords): string
     {
-        $safeStatLabel = HtmlSanitizer::safeHtmlOutput($this->getStatColumnLabel($category));
+        $safeStatLabel = HtmlSanitizer::safeHtmlOutput($this->tableRenderer->getStatColumnLabel($category));
 
         $rows = '';
         foreach ($categoryRecords as $record) {
@@ -388,67 +186,13 @@ class RecordHoldersView implements RecordHoldersViewInterface
             $rows .= '</tr>';
         }
 
-        return $this->renderCategoryTable(
+        return $this->tableRenderer->renderCategoryTable(
             $category,
             'record-table--3col-franchise',
             '<col class="col-team"><col class="col-amount"><col class="col-years">',
             '<th>Team</th><th>' . $safeStatLabel . '</th><th>Years</th>',
             $rows
         );
-    }
-
-    // ---------------------------------------------------------------
-    // Special blocks
-    // ---------------------------------------------------------------
-
-    /**
-     * Render the All-Star Appearances block.
-     *
-     * @param array{name: string, pid: int|null, teams: string, teamTids: string, amount: int, years: string} $allStar
-     */
-    private function renderAllStarBlock(array $allStar): string
-    {
-        $output = '<div class="record-category">';
-        $output .= $this->renderCategoryHeading('Most All-Star Appearances');
-        $output .= '<table class="ibl-data-table record-table ibl-table-subheading record-table--5col" data-no-responsive>';
-        $output .= '<colgroup><col class="col-player"><col class="col-team"><col class="col-amount"><col class="col-date" span="2"></colgroup>';
-        $output .= '<thead><tr><th>Player</th><th>Team</th><th>Apps</th><th colspan="2">Years</th></tr></thead>';
-        $output .= '<tbody>';
-
-        $safeName = HtmlSanitizer::safeHtmlOutput($allStar['name']);
-        $pid = $allStar['pid'];
-        $amount = (int) $allStar['amount'];
-
-        $teamLogos = '';
-        if ($allStar['teams'] !== '') {
-            $teams = explode(',', $allStar['teams']);
-            $teamTids = explode(',', $allStar['teamTids']);
-            foreach ($teams as $i => $team) {
-                $safeTid = (int) ($teamTids[$i] ?? 0);
-                $safeTeam = HtmlSanitizer::safeHtmlOutput($team);
-                $teamLogos .= '<a href="modules.php?name=Team&amp;op=team&amp;teamid=' . $safeTid . '">'
-                    . '<img src="images/topics/' . $safeTeam . '.png" alt="' . strtoupper($safeTeam) . '">'
-                    . '</a> ';
-            }
-        }
-
-        $safeYears = HtmlSanitizer::safeHtmlOutput($allStar['years']);
-        $years = $allStar['years'] !== '' ? str_replace(', ', '<br>', $safeYears) : '';
-
-        $output .= '<tr>';
-        if ($pid !== null) {
-            $output .= PlayerImageHelper::renderLargePlayerCell($pid, $allStar['name']);
-        } else {
-            $output .= '<td class="player-cell"></td>';
-        }
-        $output .= '<td>' . $teamLogos . '</td>';
-        $output .= '<td class="ibl-stat-highlight">' . $amount . '</td>';
-        $output .= '<td colspan="2">' . $years . '</td>';
-        $output .= '</tr>';
-
-        $output .= '</tbody></table></div>';
-
-        return $output;
     }
 
     // ---------------------------------------------------------------
@@ -478,70 +222,6 @@ class RecordHoldersView implements RecordHoldersViewInterface
             }
         }
         return implode(', ', $linked);
-    }
-
-    /**
-     * Render a category heading with accent left border.
-     */
-    private function renderCategoryHeading(string $category): string
-    {
-        $safeCategory = HtmlSanitizer::safeHtmlOutput($category);
-        return '<h3 class="record-category__title">' . $safeCategory . '</h3>';
-    }
-
-    /**
-     * Get the stat-specific column label for a category name.
-     *
-     * Maps category names like "Most Points in a Single Game" to abbreviations like "Pts".
-     */
-    private function getStatColumnLabel(string $category): string
-    {
-        foreach (self::STAT_LABELS as $prefix => $label) {
-            if (str_starts_with($category, $prefix)) {
-                return $label;
-            }
-        }
-
-        return 'Amount';
-    }
-
-    /**
-     * Render a single player record row.
-     *
-     * @param FormattedPlayerRecord $record
-     */
-    private function renderPlayerRecordRow(array $record, bool $multiLineAmount = false): string
-    {
-        $safeName = HtmlSanitizer::safeHtmlOutput($record['name']);
-        $safeTeam = HtmlSanitizer::safeHtmlOutput($record['teamAbbr']);
-        $safeDate = HtmlSanitizer::safeHtmlOutput($record['dateDisplay']);
-        $safeDate = str_replace("\n", '<br>', $safeDate);
-        $safeOppTeam = HtmlSanitizer::safeHtmlOutput($record['oppAbbr']);
-        $pid = $record['pid'];
-        $teamTid = $record['teamTid'];
-        $teamYr = (int) $record['teamYr'];
-        $oppTid = $record['oppTid'];
-        $oppYr = (int) $record['oppYr'];
-
-        $safeAmountRaw = HtmlSanitizer::safeHtmlOutput($record['amount']);
-        $amount = $multiLineAmount
-            ? str_replace("\n", '<br>', $safeAmountRaw)
-            : $safeAmountRaw;
-
-        $safeBoxScoreUrl = HtmlSanitizer::safeHtmlOutput($record['boxScoreUrl']);
-        $dateCell = $record['boxScoreUrl'] !== ''
-            ? '<a href="' . $safeBoxScoreUrl . '">' . $safeDate . '</a>'
-            : $safeDate;
-
-        $output = '<tr>';
-        $output .= PlayerImageHelper::renderLargePlayerCell($pid, $record['name']);
-        $output .= '<td><a href="modules.php?name=Team&amp;op=team&amp;teamid=' . $teamTid . '&amp;yr=' . $teamYr . '"><img src="images/topics/' . $safeTeam . '.png" alt="' . strtoupper($safeTeam) . '"></a></td>';
-        $output .= '<td>' . $dateCell . '</td>';
-        $output .= '<td><a href="modules.php?name=Team&amp;op=team&amp;teamid=' . $oppTid . '&amp;yr=' . $oppYr . '"><img src="images/topics/' . $safeOppTeam . '.png" alt="' . strtoupper($safeOppTeam) . '"></a></td>';
-        $output .= '<td class="ibl-stat-highlight">' . $amount . '</td>';
-        $output .= '</tr>';
-
-        return $output;
     }
 
     /**
